src/examples/recorder/requirements.txt
pyyaml==5.4.*